tests setup and fixing some tests
