+++ app/controllers/application_controller.rb
@@ -1,5 +1,8 @@
 class ApplicationController < ActionController::Base
   helper_method :current_user
+  def log_in(user)
+    session[:user_id] = user.id
+  end
   def current_user
     if session[:user_id]
       @current_user ||= User.find(session[:user_id])
