+++ app/controllers/sessions_controller.rb
@@ -1,5 +1,7 @@
 class SessionsController < ApplicationController
-  def new; end
+  def new
+    user = User.find_by_username(params[:username])
+  end
 
   def create
     user = User.find_by_username(params[:username])
