+++ app/views/events/show.html.erb
@@ -3,7 +3,7 @@
   <%= link_to @event.creator.name, @event.creator %>
 </p>
 <% if @event.attendees.include?(current_user)%>
-  <%= link_to 'Not Attending', attendance_path(@event.attendances.find_by_attendee_id  \ 
+  <%= link_to 'Not Attending', attendance_path(@event.attendances.find_by_attendee_id\
   (current_user.id)), method: :delete%>
 <%else%>
   <%= link_to 'Attend', attendances_path(event_id:@event.id), method: :post%>
