+++ spec/examples.txt
@@ -1,0 +1,3 @@
+example_id                        | status | run_time        |
+--------------------------------- | ------ | --------------- |
+./test/models/user_spec.rb[1:1:1] | passed | 0.07388 seconds |
