+++ test/application_system_test_case.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require_relative './test_helper'
 
 class ApplicationSystemTestCase < ActionDispatch::SystemTestCase
   driven_by :selenium, using: :chrome, screen_size: [1400, 1400]
