+++ test/controllers/sessions_controller_spec.rb
@@ -1,18 +1,13 @@
-require 'test_helper'
+require_relative '../test_helper'
 
 class SessionsControllerTest < ActionDispatch::IntegrationTest
   test 'should get new' do
-    get sessions_new_url
-    assert_response :success
-  end
-
-  test 'should get create' do
-    get sessions_create_url
+    get login_url
     assert_response :success
   end
 
   test 'should get destroy' do
-    get sessions_destroy_url
+    get logout_url
     assert_response :success
   end
 end
