+++ test/models/event_spec.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require_relative '../test_helper'
 
 class EventTest < ActiveSupport::TestCase
   # test "the truth" do
