+++ test/models/user_spec.rb
@@ -1,0 +1,4 @@
+require_relative '../test_helper'
+
+class UserTest < ActiveSupport::TestCase
+end
